// pops-front/pops/src/app/models/user.ts
export class User {
  id: number;
  password: string;
  firstname: string;
  name: string;
  email: string;
  role_type: string;
  is_enabled: boolean;
  date_start: Date;
  date_end: Date;
  contract: string;
  department: string;
  location: string;
  service: string;
  business_unit: string;
  got_iphone: boolean;
  manager_id: number;
  manager_name: string;

  constructor() {
  }

}